split ipc tests into three files
